backend/requirements.txt
fastapi
uvicorn
motor
pydantic>=2.0
pydantic-settings
python-dotenv
instagrapi
pyjwt
pycryptodome